Style application and start localization
