+++ src/components/results-screen/ResultsScreen.css
@@ -1,0 +1,3 @@
+.results-header {
+  color: #21C8E2;
+}
